Add distributed probe for no-TUI inference and documentation
- Implemented `distributed_probe.py` for serving and generating prompts in a distributed manner without a TUI.
- Added command-line interface for serving and generating with options for model loading, peer specification, and generation parameters.
- Created detailed documentation for the distributed inference process, including entry points, runtime flow, shard conventions, tensor transport, and usage examples.
